Move css to styles folder closes #18
