+++ src/styles/Factory.css
